% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DominanceTree.R
\name{do_dominance_tree}
\alias{do_dominance_tree}
\title{Find dominace Tree}
\usage{
do_dominance_tree(graph, nodes, from = names(head(V(graph), n = 1)),
  to = names(tail(V(graph), n = 1)))
}
\arguments{
\item{graph}{a Galois lattice of which the dominance should be found}

\item{nodes}{the labels of those nodes for which one is interested in knowing the dominace relation
for example the names of all affiliations}

\item{from}{the node from where to start the path search}

\item{to}{the node to which the shortest path should be found}
}
\value{
igraph object, a Tree describing the dominace between nodes
}
\description{
Finds the positional dominance between two nodes, by finding all shortest path between the nodes in a 
galois lattice
}
\details{
The algorithm should be used with a directed galois lattice, e.g. G <- do_galois_lattice(X, directed = TRUE).
The algorithm returns the positional dominance of the original graph, if it is applied on the
REDUCED label of the galois lattice. A Galois lattice has two possible directions and by using either of them
the positional dominance for actors and affiliations can be calculated, but once a direction is chosen
from and to nodes have to be chosen appropriately.
}
\examples{
M=matrix(c(1,1,1,0,0,0,
0,0,0,1,1,1,
1,0,0,1,0,0,
1,1,0,1,0,1),nrow=6)
colnames(M) <- c("A", "B", "C", "D")
rownames(M) <- as.character(1:6)
Galois <- do_galois_lattice(M, directed = TRUE, label = "reduced")
T <- do_dominance_tree(Galois,as.character(1:6))
plot(T)

}
\seealso{
\code{\link{do_galois_lattice}} for constructing the according input graph
}
