% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullLabelling.R
\name{do_full_label}
\alias{do_full_label}
\title{Get full Labeling of Galois Lattice}
\usage{
do_full_label(GaloisGraph, OriginalGraph)
}
\arguments{
\item{GaloisGraph}{the Galois Graph from do_galois_lattice}

\item{OriginalGraph}{the original two-mode graph used for the galois lattice}
}
\value{
igraph object, a Galois Lattice with Full Label
}
\description{
Adds to the resulting label of Galois lattice, the full label of all nodes
}
\examples{
M=matrix(c(1,1,1,0,0,0,
0,0,0,1,1,1,
1,0,0,1,0,0,
1,1,0,1,0,1),nrow=6)
colnames(M) <- c("A", "B", "C", "D")
rownames(M) <- as.character(1:6)
Galois <- do_galois_lattice(M)
Galois <- do_full_label(Galois, M)

}
\seealso{
\code{\link{do_reduced_label}} for reduced label galois lattice 
and \code{\link{galois_layout}} for correct hierarchical plots
}

