% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaloisFunction.R
\name{do_galois_lattice}
\alias{do_galois_lattice}
\title{Create a Galois lattice}
\usage{
do_galois_lattice(X, directed = FALSE, by = "best", label = "reduced",
  one_mode = FALSE)
}
\arguments{
\item{X}{a igraph object of a two mode network, or matrix}

\item{directed}{TRUE/FALSE depending on wether the output Galois lattice should be directed}

\item{by}{"col","row","best", depending if the result should be using the colnames, rownames 
or the most time efficient option}

\item{label}{"partly","full","reduced", depending if the result should have partly labeled nodes as chosen
with "by" or the full label or an reduced labeling approach}

\item{one_mode}{TRUE/FALSE if the input graph is a one-mode network}
}
\value{
igraph object, a Galois Lattice
}
\description{
Creates a Galois lattice for a two mode Graph, with labeling of chosen mapping
}
\examples{
M=matrix(c(1,1,1,0,0,0,
0,0,0,1,1,1,
1,0,0,1,0,0,
1,1,0,1,0,1),nrow=6)
colnames(M) <- c("A", "B", "C", "D")
rownames(M) <- as.character(1:6)
Galois <- do_galois_lattice(M)


}
\seealso{
\code{\link{do_full_label}} for full label galois lattice
and \code{\link{galois_layout}} for correct hierarchical plots and 
\code{\link{do_dominance_tree}} for extracting positional dominance from a galois lattice
}
