% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedLabelling.R
\name{do_reduced_label}
\alias{do_reduced_label}
\title{Get Reduced Labelling of Galois Lattice}
\usage{
do_reduced_label(GaloisGraph, OriginalGraph)
}
\arguments{
\item{GaloisGraph}{the Galois Graph from do_galois_lattice with full label}

\item{OriginalGraph}{the original two-mode graph used for the galois lattice}
}
\value{
igraph object, a Galois Lattice with Reduced Label
}
\description{
Reduces full label of Galois lattice to a specific reduced label
}
\details{
function can only be used for a full labeled galois lattice
}
\seealso{
\code{\link{galois_layout}} for correct hierarchical plots and 
\code{\link{do_dominance_tree}} for extracting positional dominance from a galois lattice
}
