% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaloisLayout.R
\name{galois_layout}
\alias{galois_layout}
\title{Layout for plotting a Galois lattice}
\usage{
galois_layout(X)
}
\arguments{
\item{X}{a Galois lattice, as the output of do_galois_lattice}
}
\value{
matrix, the layout to use in plot for the galois lattice
}
\description{
orders the nodes of a Galois lattice according to their hierarchical position
}
\examples{
M=matrix(c(1,1,1,0,0,0,
0,0,0,1,1,1,
1,0,0,1,0,0,
1,1,0,1,0,1),nrow=6)
colnames(M) <- c("A", "B", "C", "D")
rownames(M) <- as.character(1:6)
Galois <- do_galois_lattice(M)
Galois <- do_full_label(Galois, M)
Galois <- do_reduced_label(Galois, M)
plot(Galois, layout = galois_layout(Galois))


}
\seealso{
\code{\link{do_dominance_tree}} for extracting positional dominance from a galois lattice
}

