% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/southernwomenG.R
\docType{data}
\name{southernwomenG}
\alias{southernwomenG}
\title{Southern Women Data Set}
\format{An object of class \code{igraph}}
\source{
\href{https://networkdata.ics.uci.edu/netdata/html/davis.html}{networkdata}
}
\usage{
data(southernwomenG)
}
\description{
Data was collected in 1930 for ethnographic studies by Davis, Gardner and Gardner. It consists of 18 women attending 14 
events as a graph object.
}
\references{
Davis, A., Gardner, B. B. and M. R. Gardner.
Deep South. The University of Chicago Press,
Chicago, 1941.
}
\keyword{datasets}
