% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedLabelling.R
\name{do_reduced_label}
\alias{do_reduced_label}
\title{Get Reduced Labelling of Galois Lattice}
\usage{
do_reduced_label(GaloisGraph, OriginalGraph)
}
\arguments{
\item{GaloisGraph}{the Galois Graph from do_galois_lattice with full label}

\item{OriginalGraph}{the original two-mode graph used for the galois lattice}
}
\value{
igraph object, a Galois Lattice with Reduced Label
}
\description{
Reduces full label of Galois lattice to a specific reduced label
}
\details{
function can only be used for a full labeled galois lattice
}
\examples{
M=matrix(c(1,1,1,0,0,0,
0,0,0,1,1,1,
1,0,0,1,0,0,
1,1,0,1,0,1),nrow=6)
colnames(M) <- c("A", "B", "C", "D")
rownames(M) <- as.character(1:6)
Galois <- do_galois_lattice(M)
Galois <- do_full_label(Galois, M)
Galois <- do_reduced_label(Galois, M)

}
\seealso{
\code{\link{galois_layout}} for correct hierarchical plots and 
\code{\link{do_dominance_tree}} for extracting positional dominance from a galois lattice
}

